[REFACT] Review#1 수정 메뉴 정렬 알고리즘 변경

+++ model/menuModel.go
@@ -131,10 +131,10 @@ func (p *Model) SortMenu() []interface{} {
 	}
 	sort.Slice(array, func(i, j int) bool {
 		//Grade의 type이 int64이므로 type assertion을 통해 값에 접근할 때는
-		//int가 아닌 int64로 변경해주어야 합니다. 
+		//int가 아닌 int64로 변경해주어야 합니다.
 		//그렇지 않으면 firstElementI와 firstElementJ는 int의 zero value가 출력됩니다.
-		firstElementI, _ := array[i].([]interface{})[0].(int)
-		firstElementJ, _ := array[j].([]interface{})[0].(int)
+		firstElementI, _ := array[i].([]interface{})[0].(int64)
+		firstElementJ, _ := array[j].([]interface{})[0].(int64)
 		return firstElementI > firstElementJ
 	})
 	return array
